fix: int Long 변환 문제 해결

+++ src/main/java/org/example/sejonglifebe/review/dto/RatingCount.java
@@ -1,7 +1,7 @@
 package org.example.sejonglifebe.review.dto;
 
 public record RatingCount(
-        Long rating,
+        int rating,
         Long count
 ) {
 }
